% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_ids.R
\name{add_ids}
\alias{add_ids}
\title{Append new column containing DepMap IDs}
\usage{
add_ids(data, cell)
}
\arguments{
\item{data}{A tibble.}

\item{cell}{Name of column containing cell line names}
}
\value{
Provided \code{data} tibble with a new \code{depmap_id}` column
}
\description{
Matches cell line names from the given column to unique DepMap IDs. Matching
is performed using only alpha-numeric characters from the provided
names. \code{NA} will be produced when the cell line cannot be found in the
\code{\link{data_annotations}} data set or when there is
non-unique cell <-> ID mapping.
}
\examples{
data <- tibble::tibble(
  CellLine = c("MCF7", "293T", "K562"),
  viability = c(0.96, 0.58, 0.70)
)

add_ids(data, cell = "CellLine")
}
