% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotations_plot.R
\name{annotations_plot}
\alias{annotations_plot}
\title{Plot data based on given cell line feature}
\usage{
annotations_plot(data, x, y)
}
\arguments{
\item{data}{A tibble.}

\item{x}{Cell line feature}

\item{y}{Response value}
}
\value{
plot
}
\description{
Chooses the correct plot (boxplot or scatterplot) based on the type of
feature provided
}
