% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cor_expression.R
\name{cor_expression}
\alias{cor_expression}
\title{Calculate Spearman correlation between response and gene expression}
\usage{
cor_expression(data, response, ids = "depmap_id")
}
\arguments{
\item{data}{Tibble containing data}

\item{response}{Column containing response values}

\item{ids}{Column containing DepMap IDs}
}
\value{
Tibble with two columns: \code{gene} and \code{rho}
}
\description{
Calculate Spearman correlation between response and gene expression
}
\examples{
# Setup example data set
df <- tibble::tribble(
  ~"CellLine", ~"DepMapID", ~"logIC50",
  "LS513", "ACH-000007", -2.8,
  "253-J", "ACH-000011", -4.04,
  "NIH:OVCAR-3", "ACH-000001", NA
)

correlations <- cor_expression(
  data = df,
  response = "logIC50",
  ids = "DepMapID"
)
}
