% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cor_mutations.R
\name{cor_mutations}
\alias{cor_mutations}
\title{Find mutations that are associated with cell line response}
\usage{
cor_mutations(data, response, ids = "depmap_id")
}
\arguments{
\item{data}{Tibble containing data}

\item{response}{Name of column containing response data}

\item{ids}{Name of column containing depmap_ids of cell lines}
}
\value{
Tibble with effect size and significance of each gene. Contains the
columns: gene, effect (log2(mutant/wildtype)), p.value (Mann-Whitney U test),
and significance (determined using Benjamin-Hochberg correction and alpha=
0.05).
}
\description{
Find mutations that are associated with cell line response
}
