% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cor_expression.R
\name{cor_spearman}
\alias{cor_spearman}
\title{Do a spearman correlation and clean up output}
\usage{
cor_spearman(data, x, y)
}
\arguments{
\item{data}{tibble of data}

\item{x}{Name of column containing x variable}

\item{y}{Name of column containing y variable}
}
\value{
Tibble with rho and p.value columns
}
\description{
Do a spearman correlation and clean up output
}
