% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasets.R
\name{data_annotations}
\alias{data_annotations}
\title{Cell line annotations}
\usage{
data_annotations()
}
\value{
Tibble with 1,829 rows and 15 columns. Each row represents a unique
DepMap cell line. Columns included are depmap_id, cell_line_name,
stripped_cell_line_name, sex, source, sample_collection_site,
primary_or_metastasis, primary_disease, Subtype, age, lineage, lineage_subtype,
lineage_sub_subtype, lineage_molecular_subtype, and culture_type.
}
\description{
Function that returns cell line annotations. First call will download the
data from DropBox and chache it in memory for subsequent calls.
}
