% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasets.R
\name{data_drug_sensitivity}
\alias{data_drug_sensitivity}
\title{Drug sensitivity data}
\format{
A tibble with 324,205 rows and 5 columns.
\describe{
  \item{drug}{Name of drug}
  \item{depmap_id}{Unique identifier of each cell line. Formatted ACH-######}
  \item{ic50}{Concentration of drug in µM inhibits growth by 50\%}
  \item{auc}{Area under the dose response curve}
  \item{log_ic50}{log10(ic50)}
}
}
\source{
DepMap (Broad Institute): \url{https://depmap.org/portal/}
}
\usage{
data_drug_sensitivity()
}
\description{
Function that returns dose-response drug sensitivity data for 398 drugs in
975 cell lines. First call will download the data from DropBox and cache it
in memory for subsequent calls.
}
