% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasets.R
\name{data_mutations}
\alias{data_mutations}
\title{Cell line mutations}
\usage{
data_mutations()
}
\value{
Tibble with 34,365,583 rows and 3 columns. Columns are gene
<character>, depmap_id <character>, and mutant <logical>
}
\description{
Function that returns cell line annotations. First call will download the
data from DropBox and cache it in memory for subsequent calls.
}
