% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasets.R
\name{data_nutlin}
\alias{data_nutlin}
\title{Nutlin-3 data set}
\format{
A tibble with 968 rows and 6 columns.
\describe{
  \item{Cell line}{Common name of cell line}
  \item{TCGA classification}{Cancer type, annotated by The Cancer Genome Atlas}
  \item{Tissue}{Tissue of origin}
  \item{Tissue sub-type}{More detailed description of tumor type}
  \item{IC50}{Concentration of nutlin-3 (µM) required to inhibit cell growth by 50\%}
  \item{AUC}{Area under the dose response curve}
}
}
\source{
The Genomics of Drug Sensitivity in Cancer Project: \url{https://cancerrxgene.org}
}
\usage{
data_nutlin()
}
\description{
Example data set containing  sensitivity of 968 cell lines to the drug nutlin-3. Nutlin-3 is an MDM2 inhibitor,
which leads to the reactivation of p53 and cell cycle arrest.
}
