% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_expression.R
\name{exp_plot_selected}
\alias{exp_plot_selected}
\title{Generate scatterplot expression vs. response, faceted by selected gene}
\usage{
exp_plot_selected(data, selected, response, log_scale = FALSE)
}
\arguments{
\item{data}{A tibble, user data merged with expression data}

\item{selected}{Character vector with selected gene names}

\item{response}{Name of response column}

\item{log_scale}{Plot response in log-scale on y axis.}
}
\value{
Scatter plot
}
\description{
Generate scatterplot expression vs. response, faceted by selected gene
}
