% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_curve_fit.R
\name{fit_sigmoid}
\alias{fit_sigmoid}
\title{Fits data to a 4-parameter sigmoid.}
\usage{
fit_sigmoid(df, x, y)
}
\arguments{
\item{x}{Column with independent variable (e.g. dose)}

\item{y}{Column with dependent variable (e.g. viability)}

\item{data}{Tibble}
}
\value{
Nested tibble
}
\description{
Appends new columns containing the output of the \code{stats::optim} fit and
the calculated AUC, IC50, and GI50.
}
