% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_expression.R
\name{get_selected_genes}
\alias{get_selected_genes}
\title{Get list of selected genes based on row selections from table}
\usage{
get_selected_genes(data, selected)
}
\arguments{
\item{data}{Tibble of data used to make the table}

\item{selected}{list of selected row numbers}
}
\value{
Character vector of genes
}
\description{
Get list of selected genes based on row selections from table
}
