% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_help_button.R
\name{helpButton}
\alias{helpButton}
\title{Help button}
\usage{
helpButton(id, ns = NULL)
}
\arguments{
\item{id}{inputId}

\item{ns}{ns function from module}
}
\value{
actionButton with the given inputId and formatted to be extra small
with a bold question mark inside
}
\description{
Help button
}
