% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mutations.R
\name{mut_plot_selected}
\alias{mut_plot_selected}
\title{Generate boxplot genotype vs. response, faceted by selected gene}
\usage{
mut_plot_selected(data, response, log_scale = FALSE)
}
\arguments{
\item{data}{A tibble, user data merged with mutation data and filtered only for selected genes.
Contains the columns "gene" and "genotype".}

\item{response}{Name of response column}

\item{log_scale}{Plot response in log-scale on y axis.}
}
\value{
Boxplot
}
\description{
Generate boxplot genotype vs. response, faceted by selected gene
}
