% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mutations.R
\name{mut_vol_tip}
\alias{mut_vol_tip}
\title{Tooltip for the mutation volcano plot}
\usage{
mut_vol_tip(data, hover)
}
\arguments{
\item{data}{Tibble; output of gene correlation}

\item{hover}{Hover information from plot}
}
\description{
See here for reference: https://gitlab.com/-/snippets/16220
}
