% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_expression.R
\name{n_exp_matched}
\alias{n_exp_matched}
\title{Number of cell lines present in expression data set}
\usage{
n_exp_matched(data, id = "depmap_id")
}
\arguments{
\item{data}{Tibble with "depmap_id" column}

\item{id}{Column containing DepMap IDs}
}
\value{
integer, number of matching cell lines
}
\description{
Number of cell lines present in expression data set
}
