% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mutations.R
\name{n_mut_matched}
\alias{n_mut_matched}
\title{Number of cell lines present in mutation data set}
\usage{
n_mut_matched(data, id = "depmap_id")
}
\arguments{
\item{data}{A tibble with "depmap_id" column}

\item{id}{Column containing DepMap IDs}
}
\value{
integer, number of matching cell lines
}
\description{
Number of cell lines present in mutation data set
}
