% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_upload.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare uploaded data for analysis modules}
\usage{
prepare_data(data, cell, response)
}
\arguments{
\item{data}{uploaded data tibble}

\item{cell}{name of column containing cell line names}

\item{response}{name of column containing response values}
}
\value{
Tibble with only 3 columns: cell line, response, and "depmap_id"
}
\description{
Removes extra columns, averages the response values between repeated cell
line rows, and adds DepMap IDs in a new column
}
