% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mutations.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Generate volcano plot effect vs. p value}
\usage{
volcano_plot(data)
}
\arguments{
\item{data}{A tibble, output of cor_mutations()}
}
\description{
Generate volcano plot effect vs. p value
}
