% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_ids.R
\name{add_ids}
\alias{add_ids}
\title{Add column containing DepMap IDs}
\usage{
add_ids(data, name_col)
}
\arguments{
\item{data}{Tibble with one column that contains cell line names}

\item{name_col}{Name of column containing cell line names}
}
\value{
Tibble with new column appended named depmap_id
}
\description{
Matches cell lines from the given column to DepMap IDs using the
  \code{\link[cellpanelr]{annotations}} data set. Matching is performed after
  removing all non-alphanumeric characters from the provided names. \code{NA}
  will be produced when the cell line cannot be found in the 
  \code{\link[cellpanelr]{annotations}} data set or when there are non-unique
   cell-ID pairs.
}
\examples{
data <- tibble::tribble(
  ~"CellLine", ~"viability",
  "MCF7", 0.96,
  "293T", 0.58,
  "K562", 0.70
)

add_ids(data, "CellLine")
}
