% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cor_expression.R
\name{cor_expression}
\alias{cor_expression}
\title{Calculate Spearman correlation between response and gene expression}
\usage{
cor_expression(
  data,
  response = "response",
  non_coding = FALSE,
  nested_output = FALSE
)
}
\arguments{
\item{data}{Tibble}

\item{response}{character; Name of column containing response data}

\item{non_coding}{bool; Whether or not to include non-protein-coding transcripts}

\item{nested_output}{bool; Whether to include data list-column in return tibble}
}
\value{
Tibble
}
\description{
Calculate Spearman correlation between response and gene expression
}
