% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cor_expression.R
\name{cor_expression}
\alias{cor_expression}
\title{Calculate Spearman correlation between response and gene expression}
\usage{
cor_expression(df, response_col)
}
\arguments{
\item{df}{Tibble. Must have depmap_id column}

\item{response_col}{Name of column containing response data}
}
\value{
Nested tibble organized by gene_name with response vs. rna_expression
data in the "data" list-column and the rho & p.value values in the "model" 
list-column
}
\description{
Calculate Spearman correlation between response and gene expression
}
