% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cor_expression.R
\name{cor_expression}
\alias{cor_expression}
\title{Calculate Spearman correlation between response and gene expression}
\usage{
cor_expression(data, response = "response", return_nested = FALSE)
}
\arguments{
\item{data}{Tibble}

\item{response}{character; Name of column containing response data}

\item{return_nested}{bool: Whether or not the returned tibble includes the
data list-column, which holds the cell-line level expression and response
data}
}
\value{
Tibble
}
\description{
Calculate Spearman correlation between response and gene expression
}
