% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_datasets.R
\name{data_expression}
\alias{data_expression}
\title{Gene expression data}
\usage{
data_expression()
}
\value{
Tibble with 26,713561 rows and 3 columns. The columns are
\code{depmap_id}, \code{gene_name}, and \code{rna_expression} which is
expressed as \code{log2(TPM + 1)} where TPM is transcripts per million. There
are 19,177 unique genes and 1393 cell lines included in this data set.
}
\description{
Function that returns gene expression data. First call will
download the data from DropBox and cache it in memory for subsequent calls.
}
